+++ src/main/java/com/jnape/palatable/lambda/iteration/IterationInterruptedException.java
@@ -1,0 +1,17 @@
+package com.jnape.palatable.lambda.iteration;
+
+import com.jnape.palatable.lambda.functions.builtin.fn4.RateLimit;
+
+import java.util.Iterator;
+
+/**
+ * An exception thrown when a thread is interrupted while an {@link Iterator} was blocked.
+ *
+ * @see RateLimit
+ */
+public final class IterationInterruptedException extends RuntimeException {
+
+    public IterationInterruptedException(InterruptedException cause) {
+        super(cause);
+    }
+}